requirements.txt
asgiref==3.8.1
coverage==7.8.0
debugpy==1.8.16
Django==5.1.5
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
factory_boy==3.3.3
Faker==37.1.0
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.11.1
pytest-factoryboy==2.7.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
